split plugin into core and module
