--- application/Core/Plugin/Header/views/edit.phtml
@@ -1,33 +1,0 @@
-<?php /* All id's should contain the plugin id in order to prevent conflicts */ ?>
-
-<h2>Header</h2>
-
-<form id="plugin" method="post" tal:attributes="id php:'plugin'.(plugin.getId()); action php:this.pluginurl( plugin )" >
-	<input type="hidden" name="pluginaction" value="save" />
-	
-	<input type="text" name="text" tal:attributes="value header/getText" />
-	<input type="submit" value="Save" />
-</form>
-
-
-<script type="text/javascript"> 
-
-	function processJson(data) { 
-		// 'data' is the json object returned from the server 
-		alert(data.message); 
-	}
-	
-	$(document).ready(function() { 
-		$('#plugin<?php echo $ctx->plugin->getId() ?>').ajaxForm({
-			// dataType identifies the expected content type of the server response 
-			dataType:  'json', 
-	 
-			// success identifies the function to invoke when the server response 
-			// has been received 
-			success:   processJson 
-		});
-	});	
-	
-	
-	
-</script>
